requirements.txt
psycopg2-binary==2.9.3
requests==2.28.2
faker==14.1.0
celery==5.2.7